refactor: 자주 사용하는 "."을 상수로 정의
- extension 파일 분리
